create file run_review.py for all code helping execution of reviewers
